Update gradle; Make ViewModels public; Extract Settings Fragment
